Split SpellEngine in Windows and Linux specific class
...and get rid of depending on __MonoCS__ which is not defined
when building with `msbuild` (that comes with Mono 5).

Also remove test file that accidentally got moved to liblcm
although the class is was testing remained in FW.
